Add docker geode setup.sh script
This setup script will install curl, then it will loop infinitely until
it sees that the Geode REST server API is listening. Once it find that
it is open we know that the Geode server is up and running. The rest of
this setup script will PUT example keys into the exampleRegion for us to
test against.